LU-15608 sec: fix DIO for encrypted files
With Direct IO, we do not have proper page cache pages. So we need to
retrieve by ourselves the page mapping and the page index of the page
to be encrypted/decrypted.

For the index, we need to use the offset of the page within the file,
and not the object.
So we rename cl_page's cp_osc_index to cp_page_index for that purpose.
cp_osc_index is redundant with osc_async_page's oap_obj_off and only
used by osc_index(), so we also adapt this function.
cp_page_index is initialized in cl_page_alloc(), and accessed in
the OSC layer where the llcrypt primitives are called.

For the mapping, problem is page->mapping is not set to NULL on page
allocation, so it cannot safely be used to see if a page is a direct
I/O page.
Use cl_page for direct I/O and page->mapping for buffered
I/O.  (clpage->cp_inode is only set for direct I/O and
cannot easily be always set.)
Without this, we sometimes get panics when page2inode is
used in the OSC layer.  (Note the remaining use in dom is
safe because ll_dom_readpage is a page cache helper and
will never see DIO pages.)

Fixes: a71e0dd7f7 ("LU-14306 sec: get rid of bad rss-counter state messages")
Reviewed-on: https://review.whamcloud.com/46664
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/tests/sanity-sec.sh
@@ -3354,7 +3354,38 @@ test_44() {
 	cmp -bl $tmpfile $resfile ||
 		error "file $testfile is corrupted (3)"
 
-	rm -f $tmpfile $resfile
+	rm -f $tmpfile $resfile $testfile
+
+	if [ $OSTCOUNT -ge 2 ]; then
+		dd if=/dev/urandom of=$tmpfile bs=$pagesz count=1 conv=fsync
+		$LFS setstripe -S 256k -c2 $testfile
+
+		# write in file, at beginning of first stripe, buffered IO
+		dd if=$tmpfile of=$testfile bs=$pagesz count=1 \
+			conv=fsync,notrunc
+
+		# write at beginning of second stripe, direct IO
+		dd if=$tmpfile of=$testfile bs=$pagesz count=1 seek=256k \
+			oflag=seek_bytes,direct conv=fsync,notrunc
+
+		cancel_lru_locks
+
+		# read at beginning of first stripe, direct IO
+		dd if=$testfile of=$resfile bs=$pagesz count=1 \
+			iflag=direct conv=fsync
+
+		cmp -bl $tmpfile $resfile ||
+			error "file $testfile is corrupted (4)"
+
+		# read at beginning of second stripe, buffered IO
+		dd if=$testfile of=$resfile bs=$pagesz count=1 skip=256k \
+			iflag=skip_bytes conv=fsync
+
+		cmp -bl $tmpfile $resfile ||
+			error "file $testfile is corrupted (5)"
+
+		rm -f $tmpfile $resfile
+	fi
 }
 run_test 44 "encrypted file access semantics: direct IO"
